Helper,ActiveDecoratorを使いコードを見やすくまとめた
